Rename catalog_access -> catalog_adapter module
