+++ crates/integrations/datafusion/src/catalog_adapter.rs
